requirements.txt
glfw
PyOpenGL
numpy
mss
opencv-python
pyautogui
